split course resource into multiple ones
add a python script to analyze overly complex and large java files
